ai-agents-python/module1/code/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
litellm>=1.0.0
